+++ pfb/deconv/spotless.py
